Move page object into breakpoint/testing folder
